Move downscaling stack to workflows.downscaling submodule
Relocate downscale_floodmap, dilate_zsmax, apply_energy_head, the
_downscale_* helpers, and remove_disconnected_flooding from utils.py
into workflows/downscaling.py alongside the existing make_index_cog.
utils.py now retains only generic I/O and grid utilities. Expose the
submodule via workflows/__init__.py. No algorithmic changes.
