+++ hydromt_sfincs/workflows/__init__.py
@@ -4,6 +4,7 @@ from .bathymetry import *
 from .cog import *
 from .curvenumber import *
 from .discharge import *
+from .downscaling import *
 from .flwdir import *
 from .landuse import *
 from .map_overlay import *
